BREAKING: add support for react-native-gesture-handler
This adds a pager component based on [react-native-gesture-handler](kmagiera/react-native-gesture-handler), exposed as `TabViewPagerExperimental`. It'll allow you to use native animations and gestures to achieve much better performance.

This is a breaking change because the pagers receive to new animated values (`panX` and `offsetX`) to control the `position` instead of a single `position` prop, and the `position` prop in no longer an `Animated.Value`. While you can still interpolate and do other operations on it, you cannot add listeners anymore or call `.setValue` on it. If you need to listen to updates, use the position event.